Better separation of functions between files
